Shifted to OOP structure
The only reason to shift my code from procedural structure to OOP because I wanted to wrap all the methods inside one class so that they can share/send things with each other especially in case of threading, and users only have to import this class, not separate functions.
Although I haven't added any thread method but will keep it for future purposes. 

Methods Added : 
1. `LoadFromFile()`: This method loads the dictionary object stored in the `cache.txt` file (which basically contains youtube links associated with their title name -> 
{ 'crawling': <youtube_link> } and then to search that specific song name that user input, I added two different methods to get that song + link if it exists in the file.. so the first one is 
IF DictObj[songname] is True, then return the link (value associated with the key "songname")
ELSE, then perform a linear slice searching on every key of dictObj until it found.. I meant, suppose you've a string i.e., "crawling" and in that `cache.txt` file, you've a song named "crawlingonemorelight". Here, it checks in slices, means a pair of 3 characters from the string, IF all the slices (pair) of the user input string exist then return that `key` to the caller. 

2. `YoutubeSearch`: This one is simple because if the song isn't found in local directory and not even in cache file, then only choice left is to get its link from the youtube. 
In this method, I made a use of `youtube-dl` (kind of youtube downloader, but not only supports youtube :P shaddy shaddy too :) ) . Once the song information is extracted and we got the song ID, title, or whatever we are interested in, then save the {songname: <youtube link>} into `cache.txt`.. in case if user wants to play the same song and it again doesn't exist in local dir, then search through the keys in `cache.txt` because in the first run, when we did a youtube search and got the link, we stored it in `cache.txt` 
 
Haven't Added : 
1. Extracting genre (threading)
2. A good prompt for the end user to intract with methods. 
3. Might be this code is not time efficient yet (yes it isn't)