Edited SearchLocal() method
Added recursion technique to search for the songs in different directories, and if nothing is found in the first directory returned by yieldOrignalNames then it will make a second call to the generator function which returns another directory path.
Now, the same computation will be done for this path too. if the song name is found then simply checks for each piece of user input that is split into pieces by using the .split() method, and once you get the final song name that contains all the words or at least contain more than half of the words of user input (split one) then it simply breaks the while loop, otherwise it again makes a call to a generator function, ask for another directory path and do the same.

Problem: the function does the computation for the path that is already been checked.. (needs a memorization technique here)
Fix: used a dictionary to store the result that we got from each directory... if a path exists in the dictionary then it simply returns the value (tuple) associated with that key (path) and doesn't do further computation for it.
it now takes only 2 recursive calls to search a song (only case, if the song is not present in all the 3 directories) because I've only two sub-directories inside the main directory...
If a song present in the first sub-directory then it will take only 1 recursive call = main(not found) -> sub-directory(found)
If a song present in the main directory then no recursive call = main(found)

Haven't fixed : 
Exit or break condition if a song is found in any of the sub-directory, then when to do if we have True value stored in success? we don't need the control goes back to while loop and then make calls to the generator function.